+++ src/main/kotlin/com/github/imyuyu/sqltoy/dom/model/ValueElement.kt
@@ -1,0 +1,23 @@
+package com.github.imyuyu.sqltoy.dom.model
+
+import com.intellij.util.xml.*
+
+/**
+ * The interface value element.
+ *
+ * @author imyuyu
+ */
+interface ValueElement : DomElement {
+
+    /**
+     * Gets sql.
+     */
+    fun getValue(): String?
+    /**
+     * Sets value.
+     *
+     * @param content the content
+     */
+    fun setValue(content: String?)
+
+}