+++ src/main/kotlin/com/github/imyuyu/sqltoy/provider/SqlIdFindUsagesProvider.kt
@@ -10,6 +10,25 @@ import com.intellij.psi.PsiLiteralExpression
 import org.jetbrains.annotations.Nls
 import org.jetbrains.annotations.NonNls
 
-class SqlIdFindUsagesProvider : XmlFindUsagesProvider() {
+class SqlIdFindUsagesProvider : FindUsagesProvider {
 
+    override fun canFindUsagesFor(psiElement: PsiElement): Boolean {
+        return psiElement is PsiLiteralExpression
+    }
+
+    override fun getHelpId(psiElement: PsiElement): @NonNls String? {
+        return null
+    }
+
+    override fun getType(element: PsiElement): @Nls String {
+        return ""
+    }
+
+    override fun getDescriptiveName(element: PsiElement): @Nls String {
+        return ""
+    }
+
+    override fun getNodeText(element: PsiElement, useFullName: Boolean): @Nls String {
+        return "111"
+    }
 }