+++ src/main/kotlin/com/github/imyuyu/sqltoy/reference/SqlIdReferenceContributor.kt
@@ -15,7 +15,7 @@ import com.intellij.psi.xml.XmlTag
 import com.intellij.util.ProcessingContext
 import java.util.*
 
-
+@Deprecated(message = "useless")
 class SqlIdReferenceContributor: PsiReferenceContributor() {
     override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
         registrar.registerReferenceProvider(