+++ src/main/kotlin/com/github/imyuyu/sqltoy/service/EditorService.kt
@@ -1,0 +1,24 @@
+package com.github.imyuyu.sqltoy.service
+
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.ServiceManager
+import com.intellij.openapi.fileEditor.FileEditorManager
+import com.intellij.openapi.project.Project
+
+@Service(Service.Level.PROJECT)
+class EditorService(private val project: Project) {
+
+    private val fileEditorManager: FileEditorManager = FileEditorManager.getInstance(project)
+
+    companion object {
+        /**
+         * Gets instance.
+         *
+         * @param project the project
+         * @return the instance
+         */
+        fun getInstance(project: Project): EditorService {
+            return project.getService(EditorService::class.java)
+        }
+    }
+}