+++ src/main/kotlin/com/github/imyuyu/sqltoy/util/XmlUtil.kt
@@ -1,7 +1,10 @@
 package com.github.imyuyu.sqltoy.util
 
+import com.github.imyuyu.sqltoy.dom.model.SQLToy
+import com.github.imyuyu.sqltoy.indexer.SQLIdRecord
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.fileTypes.FileTypeRegistry
+import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.*
@@ -10,6 +13,7 @@ import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.xml.XmlAttribute
 import com.intellij.psi.xml.XmlFile
 import com.intellij.psi.xml.XmlTag
+import com.intellij.util.xml.DomManager
 import java.util.*
 
 
@@ -17,66 +21,26 @@ object XmlUtil {
     const val EXT: String = "sql.xml";
 
     fun findXmlPsiElement(project: Project, virtualFiles: Collection<VirtualFile>, key: String): List<PsiElement> {
-        val result: MutableList<PsiElement> = ArrayList()
+        val result: MutableList<PsiElement> = mutableListOf();
         for (virtualFile in virtualFiles) {
             val psiFile = PsiManager.getInstance(project).findFile(virtualFile)
-            if (psiFile is XmlFileImpl) {
-                val navigationElement = psiFile.navigationElement
-                if (Objects.nonNull(navigationElement)) {
-                    if (Objects.nonNull(navigationElement.children)) {
-                        val child = navigationElement.children[0]
-                        if (Objects.nonNull(child)) {
-                            if (Objects.nonNull(child.children) && child.children.size > 0) {
-                                val child1 = child.children[1]
-                                if (Objects.nonNull(child1)) {
-                                    val psiElements = child1.children
-                                    if (Objects.nonNull(psiElements) && psiElements.size > 0) {
-                                        val xmlTags: List<XmlTag> = ArrayList(
-                                            PsiTreeUtil.getChildrenOfAnyType(
-                                                child1,
-                                                XmlTag::class.java
-                                            )
-                                        )
-                                        for (xmlTag in xmlTags) {
-                                            if ("sql" == xmlTag.name) {
-                                                val xmlAttribute: XmlAttribute? = xmlTag.getAttribute("id")
-                                                val id: String? = xmlAttribute?.value
-                                                if (key == id) {
-                                                    result.add(xmlAttribute.valueElement!!)
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
+            if(psiFile !is XmlFile){
+                return result;
+            }
+            val domManager = DomManager.getDomManager(project)
+            val fileElement = domManager.getFileElement(psiFile, SQLToy::class.java) ?: return result
+
+            val sqlList = fileElement.rootElement.getSqlList()
+            for (sql in sqlList) {
+                val id = sql.getId().stringValue?.trim()
+                if (id != null && id != "") {
+                    if (key == id) {
+                        result.add(sql.getId().xmlAttribute!!)
                     }
                 }
             }
         }
         return result
-    }
-
-    fun isInjectXml(literalExpression: PsiElement, fields: MutableList<String>): Boolean {
-        if(literalExpression is PsiNewExpression && literalExpression.text.startsWith("new QueryExecutor")){
-            return true;
-        }
-        val p1 = literalExpression.parent as? PsiExpressionList ?: return false
-        val p2 = p1.parent as? PsiMethodCallExpression ?: return false
-        val text = p2.text
-        return fields.stream().filter { s ->
-            text.startsWith("$s.") || text.startsWith("super.$s.") || text.startsWith(
-                "this.$s."
-            )
-        }.findAny().isPresent
-    }
-
-    fun isNewQueryExecutor(literalExpression: PsiLiteralExpression, fields: MutableList<String>): Boolean {
-        val p1 = literalExpression.parent as? PsiExpressionList ?: return false
-        val p2 = p1.parent as? PsiNewExpression ?: return false
-        return if (!p2.text.startsWith("new QueryExecutor")) {
-            false
-        } else isInjectXml(p2, fields)
     }
 
     fun isSqltoyXml(element:PsiElement): Boolean {